IDVA5-1707 Remove an AML body
